+++ provisioning/base/general/bin/entrypoint.d/supervisord.sh
@@ -5,5 +5,5 @@
 #############################################
 
 ## Start services
-startSupervisord
+exec /opt/docker/bin/service.d/supervisor.sh