+++ test/spec/docker/php7-apache_spec.rb
@@ -11,10 +11,10 @@ describe "Dockerfile" do
     include_examples 'collection::bootstrap'
     include_examples 'collection::base'
     include_examples 'collection::base-app'
-    include_examples 'collection::php7'
+    include_examples 'collection::php7::production'
     include_examples 'collection::php-fpm7'
     include_examples 'collection::php-fpm7::local-only'
     include_examples 'collection::apache'
-    include_examples 'collection::php-fpm7::webserver-test'
+    include_examples 'collection::php-fpm7::webserver-test::production'
 
 end