+++ test/spec/shared/bootstrap/ansible.rb
@@ -1,12 +1,35 @@
 shared_examples 'bootstrap::ansible' do
 
-    it "should be ansible installed" do
-        if ['redhat', 'alpine'].include?(os[:family])
-            expect(file("/usr/bin/ansible")).to be_executable
-            expect(file("/usr/bin/ansible-playbook")).to be_executable
-        elsif ['debian', 'ubuntu'].include?(os[:family])
-            expect(file("/usr/local/bin/ansible")).to be_executable
-            expect(file("/usr/local/bin/ansible-playbook")).to be_executable
+    ansibleList = [
+        $testConfiguration[:ansiblePath] + "/ansible",
+        $testConfiguration[:ansiblePath] + "/ansible-playbook",
+        $testConfiguration[:ansiblePath] + "/ansible-galaxy",
+        $testConfiguration[:ansiblePath] + "/ansible-pull",
+        $testConfiguration[:ansiblePath] + "/ansible-doc",
+        $testConfiguration[:ansiblePath] + "/ansible-vault",
+    ].each do |file|
+        describe file("#{file}") do
+            # Type check
+            it { should be_file }
+
+            # Owner test
+            it { should be_owned_by 'root' }
+            it { should be_grouped_into 'root' }
+
+            # Read test
+            it { should be_readable.by('owner') }
+            it { should be_readable.by('group') }
+            it { should_not be_readable.by('others') }
+
+            # Write test
+            it { should be_writable.by('owner') }
+            it { should_not be_writable.by('group') }
+            it { should_not be_writable.by('others') }
+
+            # Exectuable test
+            it { should be_executable.by('owner') }
+            it { should be_executable.by('group') }
+            it { should_not be_executable.by('others') }
         end
     end