+++ test/spec/shared/samson-deployment/layout.rb
@@ -1,0 +1,101 @@
+shared_examples 'samson-deployment::layout' do
+    #########################
+    ## Directories
+    #########################
+    [
+        "/app/app",
+        "/app/tmp",
+        "/app/vendor",
+        "/app/public/assets",
+    ].each do |file|
+        describe file("#{file}") do
+            # Type check
+            it { should be_directory }
+
+            # Owner test
+            it { should be_owned_by('application') }
+            it { should be_grouped_into('application') }
+
+            # Read test
+            it { should be_readable.by('owner') }
+            it { should be_readable.by('group') }
+            it { should be_readable.by('others') }
+
+            # Write test
+            it { should be_writable.by('owner') }
+            it { should_not be_writable.by('group') }
+            it { should_not be_writable.by('others') }
+
+            # Exectuable test
+            it { should be_executable.by('owner') }
+            it { should be_executable.by('group') }
+            it { should be_executable.by('others') }
+        end
+    end
+
+    #########################
+    ## Files
+    #########################
+    [
+        "/app/public/assets/502.html",
+    ].each do |file|
+        describe file("#{file}") do
+            # Type check
+            it { should be_file }
+
+            # Owner test
+            it { should be_owned_by('application') }
+            it { should be_grouped_into('application') }
+
+            # Read test
+            it { should be_readable.by('owner') }
+            it { should be_readable.by('group') }
+            it { should be_readable.by('others') }
+
+            # Write test
+            it { should be_writable.by('owner') }
+            it { should_not be_writable.by('group') }
+            it { should_not be_writable.by('others') }
+
+            # Exectuable test
+            it { should_not be_executable.by('owner') }
+            it { should_not be_executable.by('group') }
+            it { should_not be_executable.by('others') }
+        end
+    end
+
+    #########################
+    ## Deployer
+    #########################
+
+    [
+        "/usr/local/bin/ansible",
+        "/usr/local/bin/ansible-playbook",
+        "/usr/local/bin/dep",
+        "/usr/local/bundle/bin/cap",
+    ].each do |file|
+        describe file("#{file}") do
+            # Type check
+            it { should be_file }
+
+            # Owner test
+            it { should be_owned_by('root') }
+
+            # Read test
+            it { should be_readable.by('owner') }
+            it { should be_readable.by('group') }
+            it { should be_readable.by('others') }
+
+            # Write test
+            it { should be_writable.by('owner') }
+            it { should_not be_writable.by('group') }
+            it { should_not be_writable.by('others') }
+
+            # Exectuable test
+            it { should be_executable.by('owner') }
+            it { should be_executable.by('group') }
+            it { should be_executable.by('others') }
+        end
+    end
+
+end